+++ src/run_parallel_workspaces.ts
@@ -10,10 +10,10 @@ export async function run_parallel_workspaces(cmd_a:string[], opts:monorepo_thre
 	const projects = await projects_()
 	const cmd = cmd_a.join(' ')
 	const package_name_a = projects.map(project=>project.package_name)
-	const promise_a = promise_a_<string>(projects, _promise)
+	const promise_a = promise_a_<string>(projects, promise_)
 	const stdout_a = await Promise.all(promise_a)
 	return stdout_h_package_name_(package_name_a, stdout_a)
-	async function _promise(project:project_T):Promise<string> {
+	async function promise_(project:project_T):Promise<string> {
 		const { package_dir } = project
 		return (
 			await queue.add<string>(