+++ src/run_parallel_workspaces_cli.ts
@@ -2,7 +2,7 @@ import { _param_h } from '@ctx-core/cli-args'
 import { run_parallel_workspaces } from './run_parallel_workspaces'
 const cmd_a = process.argv.slice(2)
 export async function run_parallel_workspaces_cli() {
-	const opts = _opts()
+	const opts = opts_()
 	const workspace_name_stdout =
 		await run_parallel_workspaces(cmd_a, opts)
 	for (let workspace_name in workspace_name_stdout) {
@@ -10,7 +10,7 @@ export async function run_parallel_workspaces_cli() {
 		console.info(workspace_name_stdout[workspace_name])
 	}
 }
-function _opts() {
+function opts_() {
 	const { threads, help } = _param_h(process.argv.slice(2), {
 		threads: '-t, --threads',
 		help: '-h, --help',