+++ src/sort_packages_call.d.ts
@@ -1,0 +1,2 @@
+import { pnpm_list_package_T } from './sorted_pkg_o_a_';
+export declare function sort_packages_call(fn: (pkg: pnpm_list_package_T) => Promise<void>): Promise<void>;